Move rules and matchers to common module from io.
Move RepeatRule, ByteArrayMatcher and NestedThrowableMatcher to common module.
Remove custom Digests utility.
