+++ community/common/src/test/java/org/neo4j/test/matchers/ByteArrayMatcher.java
@@ -17,12 +17,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.test;
+package org.neo4j.test.matchers;
 
 import org.hamcrest.Description;
 import org.hamcrest.TypeSafeDiagnosingMatcher;
 
-// TODO: move to common test module
 public class ByteArrayMatcher extends TypeSafeDiagnosingMatcher<byte[]>
 {
     public static ByteArrayMatcher byteArray( byte... expected )
