+++ community/kernel/src/test/java/org/neo4j/kernel/TokenCreationIT.java
@@ -34,8 +34,8 @@ import org.neo4j.graphdb.GraphDatabaseService;
 import org.neo4j.graphdb.Label;
 import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Transaction;
-import org.neo4j.test.RepeatRule;
 import org.neo4j.test.rule.EmbeddedDatabaseRule;
+import org.neo4j.test.rule.RepeatRule;
 
 /**
  * Token creation should be able to handle cases of concurrent token creation
