+++ community/neo4j-harness/LICENSES.txt
@@ -3,6 +3,7 @@ libraries. For an overview of the licenses see the NOTICE.txt file.
 
 ------------------------------------------------------------------------------
 Apache Software License, Version 2.0
+  Apache Commons Codec
   Apache Commons Compress
   Apache Commons Configuration
   Apache Commons Lang
