+++ community/neo4j-harness/NOTICE.txt
@@ -26,6 +26,7 @@ Third-party licenses
 --------------------
 
 Apache Software License, Version 2.0
+  Apache Commons Codec
   Apache Commons Compress
   Apache Commons Configuration
   Apache Commons Lang
